+++ frontend/src/api/user-config-api-service.ts
@@ -1,21 +1,20 @@
 import {Observable, of} from "rxjs";
 import ApiService, {DefaultApiService} from "./api-service";
 import {PlaylistCategories} from "../model/playlist-categories";
-import {PlaylistBouquet} from "../model/playlist-bouquet";
 
 const PLAYLIST_CATEGORIES_PATH = "user/playlist/categories";
 const PLAYLIST_BOUQUET_PATH = "user/playlist/bouquet";
 
 export default interface UserConfigApiService extends ApiService {
-    getPlaylistBouquet(): Observable<PlaylistBouquet>;
+    getPlaylistBouquet(): Observable<PlaylistCategories>;
     getPlaylistCategories(): Observable<PlaylistCategories>;
     savePlaylistBouquet(bouquet: PlaylistCategories): Observable<void>;
 }
 
 export class DefaultUserConfigApiService extends DefaultApiService implements UserConfigApiService {
 
-    getPlaylistBouquet(): Observable<PlaylistBouquet> {
-        return of(null);
+    getPlaylistBouquet(): Observable<PlaylistCategories> {
+        return this.get<PlaylistCategories>(PLAYLIST_BOUQUET_PATH);
     }
     getPlaylistCategories(): Observable<PlaylistCategories> {
         return this.get<PlaylistCategories>(PLAYLIST_CATEGORIES_PATH);