+++ frontend/src/service/user-config-service.ts
@@ -1,13 +1,12 @@
 import { Observable } from "rxjs";
 import UserConfigApiService, {DefaultUserConfigApiService} from "../api/user-config-api-service";
 import {PlaylistCategories} from "../model/playlist-categories";
-import {PlaylistBouquet} from "../model/playlist-bouquet";
 
 export default class UserConfigService {
     constructor(private userConfigApiService: UserConfigApiService = new DefaultUserConfigApiService()) {
     }
 
-    getPlaylistBouquet(): Observable<PlaylistBouquet> {
+    getPlaylistBouquet(): Observable<PlaylistCategories> {
         return this.userConfigApiService.getPlaylistBouquet();
     }
     getPlaylistCategories(): Observable<PlaylistCategories> {