+++ src/api/model/mod.rs
@@ -8,4 +8,3 @@ pub(in crate::api) mod stream_error;
 pub(crate) mod streams;
 pub(in crate::api) mod active_user_manager;
 pub(in crate::api) mod active_provider_manager;
-mod playlist_categories;